Fix some cases where you could go inside a road and that fact was not reported
Basically, in some circumstances, the road's angle was so different that
it was not even in the candidate list for intersecting objects.

To avoid this, when calculating the general rectangle used for inside of queries,
take into account the effective width ofor the current road which will prevent these events from repeating.

Also, when you aren't on a road, increase the rectangle width from 0.5 to 1,
just to be extra sure that this does not happen outside roads somewhere (not that it was seen yet).