Guard against phantom OSM changes
These changes aren't in the OSM diffs, but they can appear from the depths
of the OSMChangeIterator and wreak havoc afterwards.

It would be better to find out how that thing can materialize in the first place,
but before that's done, ignore these phantoms.

Fortunately, the OSM API will not return multiple actions for a single
object in a diff.