Use a different projection to cartesian coordinate space.
Using ECEF gave pretty okay results, but the angles did not match exactly, which
was quickly noticed by some visually impaired testers.

Because of that, we're from now on using an exact transverse mercator projection
with the zeroth longitude equal to the longitude of our default start location
as to minimize distance related errors.