Increase the default effective with for effectively inside style queries when not on a road
This fixes some artifacts where you can leave a road and a backward move
command does not return you to it.

Note that this particular value does not fix all of these, but i am not
comfortable setting the default to 50.0 which would be needed to cover all
effective widths only in Brno.