Phantom changes are not actually phantoms, they just come from a different source
Because we reuqest changes recursivly, we might get a change of a node in the
node changes and in the way changes as well.

We are doing this because we want the complete ways or relations and some parts
of these higher level objects might not lie in the specified area.

To make this a little bit more clear, degrade the phantom object warning to a trace log.

If/When we will decide that we want the objects in the area only and such risk
incomplete relations or ways, we can just stop requesting the changed objectsrecursively and the phantoms will go away completely.