Allow easier search by address
Up to now, the users were required to go through the advanced search flow
for a simple address search. Now, they can input the street and the
house number separated by a space and the query is created for them.
Note that this does not allow things like find me all the things on this street,
because we are not reliably able to determine whether the last word is a house identification or the street name.